requirements.txt
pillow
PyQt6
qt6-applications
qt6-tools
pytests